semisynthetic: simulate_reads.sh now checks for existing background read files.

+++ examples/semisynthetic/scripts/simulate_reads.sh
@@ -100,17 +100,22 @@ do
 	mkdir -p $raw_sample_dir
 	mkdir -p $trimmomatic_outdir
 
-	raw_gz1=${raw_sample_dir}/${sra_id}_1.fastq.gz
-	raw_gz2=${raw_sample_dir}/${sra_id}_2.fastq.gz
-	download_sra $sra_id $raw_gz1 $raw_gz2
-	run_trimmomatic $raw_gz1 $raw_gz2 $sra_id $trimmomatic_outdir
+	if [ -f "${tidx}_background_1.fq" ] && [ -f "${tidx}_background_2.fq" ]
+	then
+		echo "[*] Background reads already found!"
+	else
+		raw_gz1=${raw_sample_dir}/${sra_id}_1.fastq.gz
+		raw_gz2=${raw_sample_dir}/${sra_id}_2.fastq.gz
+		download_sra $sra_id $raw_gz1 $raw_gz2
+		run_trimmomatic $raw_gz1 $raw_gz2 $sra_id $trimmomatic_outdir
 
-	trimmed_1_unpaired="${trimmomatic_outdir}/${sra_id}_unmatched_1.fastq"
-	trimmed_1_paired="${trimmomatic_outdir}/${sra_id}_paired_1.fastq"
-	trimmed_2_unpaired="${trimmomatic_outdir}/${sra_id}_unmatched_2.fastq"
-	trimmed_2_paired="${trimmomatic_outdir}/${sra_id}_paired_2.fastq"
-	cat $trimmed_1_unpaired $trimmed_1_paired > ${tidx}_background_1.fq
-	cat $trimmed_2_unpaired $trimmed_2_paired > ${tidx}_background_2.fq
+		trimmed_1_unpaired="${trimmomatic_outdir}/${sra_id}_unmatched_1.fastq"
+		trimmed_1_paired="${trimmomatic_outdir}/${sra_id}_paired_1.fastq"
+		trimmed_2_unpaired="${trimmomatic_outdir}/${sra_id}_unmatched_2.fastq"
+		trimmed_2_paired="${trimmomatic_outdir}/${sra_id}_paired_2.fastq"
+		cat $trimmed_1_unpaired $trimmed_1_paired > ${tidx}_background_1.fq
+		cat $trimmed_2_unpaired $trimmed_2_paired > ${tidx}_background_2.fq
+	fi
 done < ${BACKGROUND_CSV}
